+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Utxo.hs
@@ -43,10 +43,7 @@ import qualified Cardano.Ledger.Alonzo.Rules as Alonzo (
   validateWrongNetworkInTxBody,
  )
 import Cardano.Ledger.Alonzo.Tx (AlonzoTx (..))
-import Cardano.Ledger.Alonzo.TxWits (
-  AlonzoEraTxWits (..),
-  nullRedeemers,
- )
+import Cardano.Ledger.Alonzo.TxWits (nullRedeemers)
 import Cardano.Ledger.Babbage.Collateral (collAdaBalance)
 import Cardano.Ledger.Babbage.Core
 import Cardano.Ledger.Babbage.Era (BabbageUTXO)